% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlLabeledScatter.R
\name{LabeledScatter}
\alias{LabeledScatter}
\title{rhtmlLabeledScatter HTML Widget}
\source{
https://github.com/NumbersInternational/rhtmlLabeledScatter
}
\usage{
LabeledScatter(X = NULL)
}
\arguments{
\item{param1}{is a good param}
}
\description{
A HTMLWidget that creates a labeled scatter plot.
}
\section{Usage Scenarios}{


Scenario 1: Blah blah
}
\examples{

rhtmlLabeledScatter::template('{}')

}
\author{
Po Liu <pliu0771@uni.sydney.edu.au>
}

