% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlTemplate.R
\name{template}
\alias{template}
\title{rhtmlTemplate HTML Widget}
\source{
https://github.com/NumbersInternational/rhtmlTemplate
}
\usage{
template(settingsJsonString = "{}")
}
\arguments{
\item{param1}{is a good param}
}
\description{
A HTMLWidget that ...
}
\section{Usage Scenarios}{


Scenario 1: Blah blah
}
\examples{


rhtmlTemplate::template(0.66, 400, 400, '{}')

}
\author{
First Last <first.last@gmail.com>
}

