% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlLabeledScatter.R
\name{LabeledScatter}
\alias{LabeledScatter}
\title{rhtmlLabeledScatter HTML Widget}
\source{
https://github.com/NumbersInternational/rhtmlLabeledScatter
}
\usage{
LabeledScatter(X = NULL, Y = NULL, Z = NULL, label = NULL,
  group = NULL, fixedAspect = FALSE, colors = NULL, grid = TRUE,
  origin = TRUE, y.title = "", x.title = "", z.title = "", title = "",
  x.decimals = 2, y.decimals = 2, z.decimals = 2, x.prefix = "",
  y.prefix = "", z.prefix = "", title.font.family = "Arial",
  title.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title.font.size = 16, labels.font.family = "Arial",
  labels.font.color = rgb(44, 44, 44, maxColorValue = 255),
  labels.font.size = 10, legend.show = TRUE, legend.font.color = rgb(44,
  44, 44, maxColorValue = 255), legend.font.family = "Arial",
  legend.font.size = 12, y.title.font.color = rgb(44, 44, 44, maxColorValue
  = 255), y.title.font.family = "Arial", y.title.font.size = 12,
  x.title.font.color = rgb(44, 44, 44, maxColorValue = 255),
  x.title.font.family = "Arial", x.title.font.size = 12,
  tooltip.title.font.color = rgb(44, 44, 44, maxColorValue = 255),
  tooltip.title.font.family = "Arial", tooltip.title.font.size = 10,
  width = NULL, height = NULL)
}
\arguments{
\item{param1}{is a good param}
}
\description{
A HTMLWidget that creates a labeled scatter plot.
}
\section{Usage Scenarios}{


Scenario 1: Blah blah
}
\examples{

rhtmlLabeledScatter::template('{}')

}
\author{
Po Liu <pliu0771@uni.sydney.edu.au>
}

