% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlLabeledScatter.R
\name{LabeledScatter}
\alias{LabeledScatter}
\title{rhtmlLabeledScatter HTML Widget}
\source{
https://github.com/Displayr/rhtmlLabeledScatter
}
\usage{
LabeledScatter(X = NULL, Y = NULL, Z = NULL, label = NULL,
  label.alt = NULL, group = NULL, fixed.aspect = FALSE,
  colors = c("#5B9BD5", "#ED7D31", "#A5A5A5", "#1EC000", "#4472C4", "#70AD47",
  "#255E91", "#9E480E", "#636363", "#997300", "#264478", "#43682B", "#FF2323"),
  color.transparency = NULL, grid = TRUE, origin = TRUE,
  origin.align = FALSE, y.title = "", x.title = "", z.title = "",
  title = "", x.decimals = NULL, y.decimals = NULL, z.decimals = NULL,
  x.prefix = "", y.prefix = "", z.prefix = "", x.suffix = "",
  y.suffix = "", z.suffix = "", title.font.family = "Arial",
  title.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title.font.size = 16, labels.show = TRUE, labels.font.family = "Arial",
  labels.font.color = NULL, labels.font.size = 10,
  labels.logo.scale = NULL, legend.show = TRUE,
  legend.bubbles.show = TRUE, legend.font.color = rgb(44, 44, 44,
  maxColorValue = 255), legend.font.family = "Arial", legend.font.size = 12,
  legend.bubble.font.color = rgb(44, 44, 44, maxColorValue = 255),
  legend.bubble.font.family = "Arial", legend.bubble.font.size = 10,
  legend.bubble.title.font.color = rgb(44, 44, 44, maxColorValue = 255),
  legend.bubble.title.font.family = "Arial",
  legend.bubble.title.font.size = 12, y.title.font.color = rgb(44, 44, 44,
  maxColorValue = 255), y.title.font.family = "Arial",
  y.title.font.size = 12, x.axis.show = TRUE, y.axis.show = TRUE,
  axis.font.family = "Arial", axis.font.color = "Black",
  axis.font.size = 12, x.title.font.color = rgb(44, 44, 44, maxColorValue =
  255), x.title.font.family = "Arial", x.title.font.size = 12,
  x.bounds.minimum = NULL, x.bounds.maximum = NULL,
  y.bounds.minimum = NULL, y.bounds.maximum = NULL,
  x.bounds.units.major = NULL, y.bounds.units.major = NULL,
  tooltip.title.font.color = rgb(44, 44, 44, maxColorValue = 255),
  tooltip.title.font.family = "Arial", tooltip.title.font.size = 10,
  point.radius = 2, trend.lines.show = FALSE,
  trend.lines.line.thickness = 1, trend.lines.point.size = 2,
  plot.border.show = TRUE, plot.border.color = "Black",
  plot.border.width = 1, width = NULL, height = NULL)
}
\arguments{
\item{X}{is array of x coordinates of data set}

\item{Y}{is array of y coordinates of data set}

\item{Z}{is array of magnitudes for each set of x,y coordinates (for bubble charts). This is optional}

\item{label}{is the array of text labels for the data set (can supply an url to show logos)}

\item{label.alt}{is an optional array of alternate label text when an url was provided as the label. NOTE: must be same length as label}

\item{group}{is the array of group name for each data point}

\item{fixed.aspect}{Default to FALSE. Cannot be guarenteed if any of the axis bounds are set.}

\item{colors}{is the color wheel to be used when plotting the data points. Defaults to Q color wheel.}

\item{color.transparency}{Value 0-1 specifying the transparency level of the plot points. Defaults to 1 without Z and 0.8 with Z}

\item{grid}{Defaults to TRUE. Shows the grid lines.}

\item{origin}{Defaults to FALSE. Shows the origin lines as dotted if not along axis.}

\item{origin.align}{Defaults to FALSE. Aligns the origin lines as closely to axis as possible.}

\item{y.title}{is the title text given to the y axis}

\item{x.title}{is the title text given to the x axis}

\item{z.title}{is the title text given to the bubble size}

\item{x.decimals}{the number of decimals in the x axis}

\item{y.decimals}{the number of decimals in the y axis}

\item{z.decimals}{the number of decimals in the bubble size axis}

\item{x.prefix}{A string that prefixes all x values(eg. "$")}

\item{y.prefix}{A string that prefixes all y values(eg. "$")}

\item{z.prefix}{A string that prefixes all bubble values(eg. "$")}

\item{x.suffix}{A string that suffixes all x values(eg. "kg")}

\item{y.suffix}{A string that suffixes all y values(eg. "kg")}

\item{z.suffix}{A string that suffixes all bubble values(eg. "kg")}

\item{title.font.family}{is the font family of the plot title}

\item{title.font.color}{is the font color of the plot title}

\item{labels.show}{Toggle for showing labels. Defaults to true if labels array given}

\item{labels.font.family}{is the font family of the labels}

\item{labels.font.color}{is the font color of the labels. NOTE: This overrides the color if it is set}

\item{labels.font.size}{is the font size of the labels}

\item{labels.logo.scale}{is a vector of scaling factors for label logos}

\item{legend.show}{is the toggle to show the legend. Defaults to TRUE}

\item{legend.bubbles.show}{toggle to show the bubble sizes in the legend. Defaults to TRUE}

\item{legend.font.color}{is the font color of the legend.}

\item{legend.font.family}{is the font family of the legend}

\item{legend.font.size}{is the font size of the legend}

\item{legend.bubble.font.color}{is the font color of the legend bubble values.}

\item{legend.bubble.font.family}{is the font family of the legend bubble values}

\item{legend.bubble.font.size}{is the font size of the legend bubble values}

\item{legend.bubble.title.font.color}{is the font color of the legend bubble title.}

\item{legend.bubble.title.font.family}{is the font family of the legend bubble title}

\item{legend.bubble.title.font.size}{is the font size of the legend bubble title}

\item{y.title.font.color}{is the font color of the y axis title}

\item{y.title.font.family}{is the font family of the y axis title}

\item{y.title.font.size}{is the font size of the y axis title}

\item{x.axis.show}{Boolean toggle to show the x axis tick markers (Default is TRUE).}

\item{y.axis.show}{Boolean toggle to show the y axis tick markers (Default is TRUE).}

\item{axis.font.family}{Font Family of the axis labels}

\item{axis.font.color}{Font color of the axis labels}

\item{axis.font.size}{Font size of the axis labels}

\item{x.title.font.color}{is the font color of the x axis title}

\item{x.title.font.family}{is the font family of the x axis title}

\item{x.title.font.size}{is the font size of the x axis title}

\item{x.bounds.minimum}{Integer or NULL; set minimum of range for plotting on the x axis}

\item{x.bounds.maximum}{Integer or NULL; set minimum of range for plotting on the x axis}

\item{y.bounds.minimum}{Integer or NULL; set minimum of range for plotting on the y axis}

\item{y.bounds.maximum}{Integer or NULL; set minimum of range for plotting on the y axis}

\item{x.bounds.units.major}{Integer or NULL; set the distance between each tick mark on the x axis.}

\item{y.bounds.units.major}{Integer or NULL; set the distance between each tick mark on the y axis.}

\item{point.radius}{Radius of the points when bubble (Z) parameter not supplied. Defaults to 2.}

\item{trend.lines.show}{Boolean toggle to show trendlines based on groups given}

\item{trend.lines.line.thickness}{An integer for the thickness of the trendlines (Default is 1px)}

\item{trend.lines.point.size}{An integer to set the size of the data points when a trendline is drawn. This setting overrides Z sizes.}

\item{plot.border.show}{Boolean toggle to show border around plot area (Default is TRUE).}

\item{plot.border.color}{Color of border around plot area (Default is black).}

\item{plot.border.width}{Width of border around plot area in px (Default is 1).}

\item{width}{is the width of the plot. Defaults to max of window}

\item{height}{is the height of the plot. Defaults to the max of window}

\item{title.font}{size is the font size of the plot title}

\item{tooltip.font.color}{is the font color of the tooltips}

\item{tooltip.font.family}{is the font family of the tooltips}

\item{tooltip.font.size}{is the font size of the tooltips}
}
\description{
A HTMLWidget that creates a labeled scatter plot.
}
\author{
Po Liu <po.liu@displayr.com>
}
