% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlLabeledScatter.R
\name{LabeledScatter}
\alias{LabeledScatter}
\title{Labeled scatterplot HTMLWidget}
\source{
https://github.com/Displayr/rhtmlLabeledScatter
}
\usage{
LabeledScatter(X = NULL, Y = NULL, Z = NULL,
  axis.font.color = "Black", axis.font.family = "Arial",
  axis.font.size = 12, color.transparency = NULL,
  colors = c("#5B9BD5", "#ED7D31", "#A5A5A5", "#1EC000", "#4472C4",
  "#70AD47", "#255E91", "#9E480E", "#636363", "#997300", "#264478",
  "#43682B", "#FF2323"), debug.mode = FALSE, fixed.aspect = FALSE,
  footer = "", footer.font.color = rgb(44, 44, 44, maxColorValue =
  255), footer.font.family = "Arial", footer.font.size = 8,
  grid = TRUE, group = NULL, height = NULL, label = NULL,
  label.alt = NULL, label.placement.maxAngle = 2 * 3.1415,
  label.placement.maxMove = 5, label.placement.numSweeps = 500,
  label.placement.seed = 1, label.placement.temperature.initial = 0.01,
  label.placement.temperature.final = 1e-04,
  label.placement.weight.distance = 10,
  label.placement.weight.distance.multiplier.centeredAboveAnchor = 1,
  label.placement.weight.distance.multiplier.centeredUnderneathAnchor = 1.5,
  label.placement.weight.distance.multiplier.besideAnchor = 4,
  label.placement.weight.distance.multiplier.diagonalOfAnchor = 15,
  label.placement.weight.labelAncOverlap = 8,
  label.placement.weight.labelLabelOverlap = 12,
  labels.font.color = NULL, labels.font.family = "Arial",
  labels.font.size = 10, labels.logo.scale = NULL,
  labels.show = TRUE, leaderline.distance.minimum = 10,
  leaderline.distance.nearbyAnchors = 10,
  legend.bubble.font.color = rgb(44, 44, 44, maxColorValue = 255),
  legend.bubble.font.family = "Arial", legend.bubble.font.size = 10,
  legend.bubble.title.font.color = rgb(44, 44, 44, maxColorValue = 255),
  legend.bubble.title.font.family = "Arial",
  legend.bubble.title.font.size = 12, legend.bubbles.show = TRUE,
  legend.font.color = rgb(44, 44, 44, maxColorValue = 255),
  legend.font.family = "Arial", legend.font.size = 12,
  legend.show = TRUE, origin = TRUE, origin.align = FALSE,
  plot.border.color = "Black", plot.border.show = TRUE,
  plot.border.width = 1, point.radius = if (is.null(Z)) 2 else 4,
  subtitle = "", subtitle.font.color = rgb(44, 44, 44, maxColorValue =
  255), subtitle.font.family = "Arial", subtitle.font.size = 12,
  title = "", title.font.color = rgb(44, 44, 44, maxColorValue = 255),
  title.font.family = "Arial", title.font.size = 16,
  tooltip.font.color = rgb(44, 44, 44, maxColorValue = 255),
  tooltip.font.family = "Arial", tooltip.font.size = 10,
  tooltip.text = NULL, trend.lines.line.thickness = 1,
  trend.lines.point.size = 2, trend.lines.show = FALSE, width = NULL,
  x.axis.show = TRUE, x.bounds.maximum = NULL,
  x.bounds.minimum = NULL, x.bounds.units.major = NULL,
  x.decimals = NULL, x.format = NULL, x.levels = NULL,
  x.prefix = "", x.suffix = "", x.title = "",
  x.title.font.color = rgb(44, 44, 44, maxColorValue = 255),
  x.title.font.family = "Arial", x.title.font.size = 12,
  y.axis.show = TRUE, y.bounds.maximum = NULL,
  y.bounds.minimum = NULL, y.bounds.units.major = NULL,
  y.decimals = NULL, y.format = NULL, y.levels = NULL,
  y.prefix = "", y.suffix = "", y.title = "",
  y.title.font.color = rgb(44, 44, 44, maxColorValue = 255),
  y.title.font.family = "Arial", y.title.font.size = 12,
  z.decimals = NULL, z.prefix = "", z.suffix = "", z.title = "")
}
\arguments{
\item{X}{is a vector of x coordinates of data set}

\item{Y}{is vector of y coordinates of data set}

\item{Z}{is vector of magnitudes for each set of x,y coordinates (for bubble charts). This is optional}

\item{axis.font.color}{Font color of the axis labels}

\item{axis.font.family}{Font Family of the axis labels}

\item{axis.font.size}{Font size of the axis labels}

\item{color.transparency}{Value 0-1 specifying the transparency level of the plot points. Defaults to 1 without Z and 0.8 with Z}

\item{colors}{is the color wheel to be used when plotting the data points. Defaults to Q color wheel.}

\item{debug.mode}{Boolean toggle to display widget internals for debugging (Default is FALSE)}

\item{fixed.aspect}{Default to FALSE. Cannot be guarenteed if any of the axis bounds are set.}

\item{footer}{is the footer text given at the bottom at the plot}

\item{footer.font.color}{is the font color of the footer text}

\item{footer.font.family}{is the font of the footer text}

\item{footer.font.size}{is the font size of the footer text}

\item{grid}{Defaults to TRUE. Shows the grid lines.}

\item{group}{is the array of group name for each data point}

\item{height}{is the height of the plot. Defaults to the max of window}

\item{label}{is the array of text labels for the data set (can supply an url to show logos)}

\item{label.alt}{is an optional array of alternate label text when an url was provided as the label. NOTE: must be same length as label}

\item{label.placement.maxAngle}{Label placement algorithm setting to determine the domain of angles chosen for mcrotate (Default is 2*Pi).}

\item{label.placement.maxMove}{Label placement algorithm setting to determine how far in pixels a move is made (Default is 5).}

\item{label.placement.numSweeps}{Label placement algorithm number of sweeps through the dataset (Default is 500).}

\item{label.placement.seed}{Label placement algorithm setting for the randomiser seed (Default is 1).}

\item{label.placement.temperature.initial}{Label placement algorithm initial temperature (Default is 0.01).}

\item{label.placement.temperature.final}{Label placement algorithm final temperature (Default is 0.0001).}

\item{label.placement.weight.distance}{Label placement algorithm weight for the distance between the label and the point (Default is 10.0)}

\item{label.placement.weight.distance.multiplier.centeredAboveAnchor}{TODO document}

\item{label.placement.weight.distance.multiplier.centeredUnderneathAnchor}{TODO document}

\item{label.placement.weight.distance.multiplier.besideAnchor}{TODO document}

\item{label.placement.weight.distance.multiplier.diagonalOfAnchor}{TODO document}

\item{label.placement.weight.labelAncOverlap}{Label placement algorithm weight fo the overlap between the point and label (Default is 8.0)}

\item{label.placement.weight.labelLabelOverlap}{Label placement algorithm weight for the overlap between two labels (Default is 12.0)}

\item{labels.font.color}{is the font color of the labels. NOTE: This overrides the color if it is set}

\item{labels.font.family}{is the font family of the labels}

\item{labels.font.size}{is the font size of the labels}

\item{labels.logo.scale}{is a vector of scaling factors for label logos}

\item{labels.show}{Toggle for showing labels. Defaults to true if labels array given}

\item{leaderline.distance.minimum}{Distance between anchor and label must meet or exceed this value before a leader line will be drawn (Default is 10)}

\item{leaderline.distance.nearbyAnchors}{If a label has other anchors nearby, a leader line be added regardless of distance between anchor and label (Default is 10)}

\item{legend.bubble.font.color}{is the font color of the legend bubble values.}

\item{legend.bubble.font.family}{is the font family of the legend bubble values}

\item{legend.bubble.font.size}{is the font size of the legend bubble values}

\item{legend.bubble.title.font.color}{is the font color of the legend bubble title.}

\item{legend.bubble.title.font.family}{is the font family of the legend bubble title}

\item{legend.bubble.title.font.size}{is the font size of the legend bubble title}

\item{legend.bubbles.show}{toggle to show the bubble sizes in the legend. Defaults to TRUE}

\item{legend.font.color}{is the font color of the legend.}

\item{legend.font.family}{is the font family of the legend}

\item{legend.font.size}{is the font size of the legend}

\item{legend.show}{is the toggle to show the legend. Defaults to TRUE}

\item{origin}{Defaults to FALSE. Shows the origin lines as dotted if not along axis.}

\item{origin.align}{Defaults to FALSE. Aligns the origin lines as closely to axis as possible.}

\item{plot.border.color}{Color of border around plot area (Default is black).}

\item{plot.border.show}{Boolean toggle to show border around plot area (Default is TRUE).}

\item{plot.border.width}{Width of border around plot area in px (Default is 1).}

\item{point.radius}{Radius of the points when bubble parameter \code{Z} is not supplied. Defaults to 2.
When the \code{Z} is supplied, the points are scaled so that the largest point has a radius of
\code{point.radius * 50/3} (i.e. a diameter of roughly an inch for the default value).}

\item{subtitle}{is the subtitle text given to the plot}

\item{subtitle.font.color}{is the font color of the subtitle text}

\item{subtitle.font.family}{is the font of the subtitle text}

\item{subtitle.font.size}{is the font size of the subtitle text}

\item{title}{is the title text given to the plot}

\item{title.font.color}{is the font color of the plot title}

\item{title.font.family}{is the font family of the plot title}

\item{title.font.size}{is the font size of the plot title}

\item{tooltip.font.color}{is the font color of the tooltips}

\item{tooltip.font.family}{is the font family of the tooltips}

\item{tooltip.font.size}{is the font size of the tooltips}

\item{tooltip.text}{is an array of text containing custom tool tip text that appears on mouse hover ('\\n' for new line)}

\item{trend.lines.line.thickness}{An integer for the thickness of the trendlines (Default is 1px)}

\item{trend.lines.point.size}{An integer to set the size of the data points when a trendline is drawn. This setting overrides Z sizes.}

\item{trend.lines.show}{Boolean toggle to show trendlines based on groups given}

\item{width}{is the width of the plot. Defaults to max of window}

\item{x.axis.show}{Boolean toggle to show the x axis tick markers (Default is TRUE).}

\item{x.bounds.maximum}{Integer or NULL; set minimum of range for plotting on the x axis}

\item{x.bounds.minimum}{Integer or NULL; set minimum of range for plotting on the x axis}

\item{x.bounds.units.major}{Integer or NULL; set the distance between each tick mark on the x axis.}

\item{x.decimals}{the number of decimals in the x axis}

\item{x.format}{A string that is interpreted for the format of the x axis labels. Default is NULL.}

\item{x.levels}{is the levels for the categorical X input array. Default is levels(X)}

\item{x.prefix}{A string that prefixes all x values(eg. "$")}

\item{x.suffix}{A string that suffixes all x values(eg. "kg")}

\item{x.title}{is the title text given to the x axis}

\item{x.title.font.color}{is the font color of the x axis title}

\item{x.title.font.family}{is the font family of the x axis title}

\item{x.title.font.size}{is the font size of the x axis title}

\item{y.axis.show}{Boolean toggle to show the y axis tick markers (Default is TRUE).}

\item{y.bounds.maximum}{Integer or NULL; set minimum of range for plotting on the y axis}

\item{y.bounds.minimum}{Integer or NULL; set minimum of range for plotting on the y axis}

\item{y.bounds.units.major}{Integer or NULL; set the distance between each tick mark on the y axis.}

\item{y.decimals}{the number of decimals in the y axis}

\item{y.format}{A string that is interpreted for the format of the y axis labels. Default is NULL.}

\item{y.levels}{is the levels for the categorical Y input array. Default is levels(Y)}

\item{y.prefix}{A string that prefixes all y values(eg. "$")}

\item{y.suffix}{A string that suffixes all y values(eg. "kg")}

\item{y.title}{is the title text given to the y axis}

\item{y.title.font.color}{is the font color of the y axis title}

\item{y.title.font.family}{is the font family of the y axis title}

\item{y.title.font.size}{is the font size of the y axis title}

\item{z.decimals}{the number of decimals in the bubble size axis}

\item{z.prefix}{A string that prefixes all bubble values(eg. "$")}

\item{z.suffix}{A string that suffixes all bubble values(eg. "kg")}

\item{z.title}{is the title text given to the bubble size}
}
\description{
A HTMLWidget that creates a labeled scatter plot.
}
\author{
Po Liu <po.liu@displayr.com>
}
